Split carousels into Scroll and Button types
