--- Assets/Prefabs/Carousel.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 69d0830620905644593c0597b8429ad1
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
