+++ Assets/Scripts/UI/CoreGameplay/Carousel/Carousel_Button.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6fc61d30fbeb469489e02094b30fc5f7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
