+++ Assets/Scripts/UI/CoreGameplay/Carousel/Carousel_Scroll.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: df7a1e1d7d39f7443a7ddf3291878f64
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
